backend/requirements.txt
fastapi
pydantic
uvicorn
langchain
langchain_experimental
langchain-community
langchain-core
langchain-google-genai
python-dotenv
tiktoken
faiss-cpu
streamlit
protobuf
mysql-connector-python
sentence-transformers
chromadb
PyPDF2
python-multipart
google-generativeai
groq
langchain-groq
langchain-pinecone
beautifulsoup4
httptools
pip-chill
websockets
watchfiles